# infra/openapi/system-api.yaml
openapi: 3.0.3
info:
  title: Livskin System API
  version: 1.0.0
  description: |
    API uniforme expuesta por los 3 VPS de Livskin para observabilidad
    cross-VPS y operación AI-friendly.

    **Implementación:**
    - VPS 1 (livskin-wp): `livskin-sensor` systemd service en puerto 9100
    - VPS 2 (livskin-ops): container `livskin-sensor` en puerto 9100
    - VPS 3 (livskin-erp): endpoints en `erp-flask` (puerto 8000) — los
      mismos paths con prefix `/api/internal/`

    **Auth:**
    - `/api/health` — público (liveness solamente)
    - `/api/system-state` — header `X-Internal-Token`
    - `/api/internal/audit-event` — header `X-Internal-Token`
    - `/api/system-map.json` — público (sin secretos)

    **Whitelist UFW:** puerto 9100 solo accesible desde DO VPC `10.114.0.0/20`.

  contact:
    name: Livskin Ops
    url: https://github.com/DarioUrrutia/livskin-revops
servers:
  - url: http://10.114.0.3:9100
    description: VPS 1 (livskin-wp) sensor
  - url: http://10.114.0.2:9100
    description: VPS 2 (livskin-ops) sensor
  - url: https://erp.livskin.site
    description: VPS 3 (livskin-erp) — endpoints en erp-flask con prefix /api/internal/
components:
  securitySchemes:
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  schemas:
...
    AuditEvent:
      type: object
      required: [action]
      properties:
        action:
          type: string
          example: infra.deploy_completed
          description: Nombre del evento — debe estar en KNOWN_ACTIONS del audit_service o se loguea como warning
        entity_type:
          type: string
          example: vps
        entity_id:
          type: string
          example: livskin-vps-erp
        result:
          type: string
          enum: [success, failure]
          default: success
        error_detail:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
paths:
...
  /api/internal/audit-event:
    post:
      summary: Recibe eventos de infra (deploys, backups, alertas)
      description: |
        Solo expuesto en VPS 3 (erp-flask). Los workflows GHA y crons
        de los 3 VPS POST aquí para persistir eventos en audit_log.
      operationId: postAuditEvent
      security:
        - InternalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuditEvent' }
      responses:
        '201':
          description: Evento registrado
        '400':
          description: Body inválido (action faltante)
        '403':
          description: X-Internal-Token inválido
...